fzdm_one_piece/requirements.txt
bs4==0.0.1
faker==4.0.2
requests==2.21.0
pillow==6.2.1
